Move VerificationBadge to core
