+++ front/app/containers/MainHeader/UserMenu/VerificationBadge.tsx
